Extract PumpCommHandler and BleChangeReceiver from CommService
Continue the Phase 0 refactoring by extracting the remaining inner classes
from CommService into standalone classes in the pump/ package:

- Extract PumpCommHandler (~750 lines) to pump/PumpCommHandler.kt with its
  nested Pump inner class, now accepting CommServiceCallbacks instead of
  implicit this@CommService references
- Extract BleChangeReceiver and BondState enum to pump/BleChangeReceiver.kt
- Move historyLogFetcher, historyLogSyncWorker, and related methods into
  PumpCommHandler where they logically belong
- Add getWearPrefs() and sendPumpCommMessages() to CommServiceCallbacks
  interface for PumpCommHandler's bolus signature validation and command
  routing needs
- CommService reduced from ~1780 to ~900 lines, now primarily a thin
  wiring layer that implements CommServiceCallbacks

No behavioral changes — pure structural refactoring for testability.
